+++ packages/compiler-cli/src/ngcc/BUILD.bazel
@@ -16,5 +16,6 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/host",
         "//packages/compiler-cli/src/ngtsc/metadata",
         "//packages/compiler-cli/src/ngtsc/transform",
+        "//packages/compiler-cli/src/ngtsc/translator",
     ],
 )
