+++ packages/compiler-cli/src/ngtsc/translator/src/translator.ts
